fix(checkstyle): fix checkstyle warnings
- integrate newer checkstyle version into gradle -> change checkstyle.xml due to new xml
- fix a few other warnings in other classes as well due to new checkstyle version

+++ src/main/java/com/box/sdk/LargeFileUpload.java
@@ -180,7 +180,7 @@ public final class LargeFileUpload {
                 }
             }
             if (this.executorService.getQueue().size() < maxQueueSize) {
-                long diff = fileSize - (long) processed;
+                long diff = fileSize - processed;
                 //The size last part of the file can be lesser than the part size.
                 if (diff < (long) partSize) {
                     partSize = (int) diff;
